radiotap: fix parsing of namespaces in presence bitmaps
Previously the parser didn't handle vendor namespaces or repeated
fields using the radiotap namespace.  Enhance the presence bitmap
parser in order to capture this data.  Currently we only report
vendor namespace data as a big block of data and a presence bitmap.

In order to capture repeated fields, the parser has a new option,
valuelist, which will return the fields as a list when encountered,
rather than as a dictionary.  Thus a header that has, for example,
two TSFT fields will have each reported separately.  In dictionary
mode, only the latest field is returned.

Signed-off-by: Bob Copeland <me@bobcopeland.com>